Zero steps UI empty state
