+++ packages/client/src/components/SingleTestResultPage/styles.module.css
@@ -21,7 +21,7 @@
   left: 0;
   right: 0;
   bottom: 0;
-  background: url('./globalAssets/loader.svg') no-repeat center;
+  background: url('../../globalAssets/loader.svg') no-repeat center;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -71,7 +71,7 @@
 }
 
 .arrowIconSvg {
-  background-image: url('./globalAssets/arrow.svg');
+  background-image: url('./../../globalAssets/arrow.svg');
   width: 8px;
   height: 12px;
 }
