+++ packages/client/src/components/ZeroStepsTest/styles.module.css
@@ -1,0 +1,57 @@
+.screenshot {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  margin: 0 3px;
+  border-radius: 5px;
+  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.11);
+}
+
+.screenshotInner {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.screenshot img {
+  height: 100%;
+  width: 100%;
+  border-radius: 5px;
+}
+
+.screenshotHighlight {
+  display: inline-block;
+  position: absolute;
+  box-shadow: 0px 0px 0px 2000px rgba(0, 0, 0, 0.4);
+}
+
+.highlightRect {
+  outline: 2px solid #70d5ff;
+  /* box-shadow: 0 0 4px 4px #c734d1; */
+}
+
+.emptyStateWrapper {
+  display: flex;
+  flex-direction: column;
+  box-shadow: 0 3px 7px 0 rgba(0, 0, 0, 0.1);
+  border-radius: 5px;
+  overflow: hidden;
+  background-color: #ffffff;
+  margin: 7px;
+  height: 100%;
+  justify-content: center;
+}
+
+.emptyStateWrapper img {
+  margin-bottom: 48px;
+}
+
+.emptyStateWrapper h3 {
+  text-align: center;
+  color: #7a838e;
+  font-size: 16px;
+  font-weight: 600;
+  letter-spacing: 0;
+  line-height: 14px;
+  font-family: 'Mont-Regular', sans-serif;
+}
