+++ packages/jest-tester-and-example/for-integration-test/example1.test.ts
@@ -1,9 +1,9 @@
 describe('Some test', () => {
   it('First Test pass', async () => {
-    await page.goto('http://jsbin.testim.io/tog');
-    await page.click('#forwarning');
-    await expect(page).toMatchElement('#forlog', { text: 'Make log' });
-    await page.click('#forlog');
+    // await page.goto('http://jsbin.testim.io/tog');
+    // await page.click('#forwarning');
+    // await expect(page).toMatchElement('#forlog', { text: 'Make log' });
+    // await page.click('#forlog');
   }, 10_000);
 
   it('Test that should fail', async () => {
